Finally got rid of the ActiveX control. Now using LibVlcWrapper from Codeproject which is very complex but works. I put it on gitignore as it is large and I will never touch it.

+++ ThePlayer (WF)/Mainform.cs
@@ -31,7 +31,7 @@ namespace ThePlayer
             }
         }
 
-        void ActivePlayer_PositionChanged(int position)
+        void ActivePlayer_PositionChanged(double position)
         {
             this.Invoke((MethodInvoker)delegate
             {
